check ssh host aliases when no explicit github url
I add use ssh host aliases i.e:

Host gh
    HostName github.com
    User git

This means I can reference remotes in the form gh:user/repo.
Unfortunately, git-pr is currently blind to that. But no longer!

One not so nice thing is that Net::SSH::Perl::Config forces you to
reinstantiate its object (and reread the file) to look for a different
host. I'm going to see if I can patch that package to make it a bit
nicer for us in this regard. However, it's still quite quick - and this
code only runs if a "normal" github.com url is not found.